refactor: auth zustand 컨벤션 적용
