--- src/feature/auth/store/auth.store.ts
@@ -1,7 +1,0 @@
-import { create } from 'zustand';
-import { AuthState } from './auth.type';
-import { createAuthSlice } from './auth.slice';
-
-export const useAuthStore = create<AuthState>()((...args) => ({
-  ...createAuthSlice(...args),
-}));
